boards/stm32l4: rename common clock config header
