+++ boards/nucleo-l476rg/include/periph_conf.h
@@ -24,7 +24,7 @@
 #define PERIPH_CONF_H
 
 #include "periph_cpu.h"
-#include "l4/cfg_clock_80_1.h"
+#include "l4/cfg_clock_default.h"
 #include "cfg_i2c1_pb8_pb9.h"
 #include "cfg_rtt_default.h"
