+++ boards/p-l496g-cell02/include/periph_conf.h
@@ -20,7 +20,7 @@
 #define PERIPH_CONF_H
 
 #include "periph_cpu.h"
-#include "l4/cfg_clock_80_1.h"
+#include "l4/cfg_clock_default.h"
 #include "cfg_rtt_default.h"
 #include "cfg_timer_tim2.h"
